# Conflicts:
#	warehouse-service/src/main/java/sum25/group03/warehouseservice/entity/Instrument.java
#	warehouse-service/src/main/java/sum25/group03/warehouseservice/repository/SpecificConfigRepo.java
#	warehouse-service/src/main/java/sum25/group03/warehouseservice/service/config/ConfigService.java
#	warehouse-service/src/main/java/sum25/group03/warehouseservice/service/config/ConfigServiceImpl.java
#	warehouse-service/src/main/java/sum25/group03/warehouseservice/service/instrument/InstrumentServiceImpl.java

+++ instrument-service/src/main/java/sum25/group03/instrumentservice/model/RawTestResult.java
@@ -18,10 +18,10 @@ public class RawTestResult {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ResultID")
-    private Integer resultId;
+    private Long resultId;
 
     @Column(name = "TestOrderID")
-    private Integer testOrderId;
+    private Long testOrderId;
 
     @ManyToOne
     @JoinColumn(name = "InstrumentID")
